+++ src/interfaces/query-status.ts
@@ -1,4 +1,0 @@
-export interface IQueryStatus {
-  message: string
-  status: string
-}